+++ apps/prairielearn/src/batched-migrations/20240108185602_group_users__group_config_id__backfill.sql
@@ -1,0 +1,10 @@
+-- BLOCK update_group_users_group_config_id
+UPDATE group_users AS gu
+SET
+  group_config_id = g.group_config_id
+FROM
+  groups AS g
+WHERE
+  gu.group_id = g.id
+  AND gu.group_id >= $min
+  AND gu.group_id <= $max;