+++ apps/prairielearn/src/pages/instructorCourseAdminStaff/instructorCourseAdminStaff.js
@@ -1,5 +1,4 @@
-import { selectCourseInstancesWithStaffAccess } from '../../models/course-instances';
-
+//@ts-check
 const asyncHandler = require('express-async-handler');
 const express = require('express');
 const router = express.Router();
@@ -10,6 +9,7 @@ const { logger } = require('@prairielearn/logger');
 const error = require('@prairielearn/error');
 const sqldb = require('@prairielearn/postgres');
 const { idsEqual } = require('../../lib/id');
+const { selectCourseInstancesWithStaffAccess } = require('../../models/course-instances');
 
 const path = require('path');
 const debug = require('debug')('prairielearn:' + path.basename(__filename, '.js'));
@@ -100,6 +100,9 @@ router.post(
       const result = await async.reduce(
         [...uids],
         { given_cp: [], not_given_cp: [], not_given_cip: [], errors: [] },
+        /**
+         * @param {{ given_cp: string[], not_given_cp: string[], not_given_cip: string[], errors: string[] }} memo
+         */
         async (memo, uid) => {
           let result;
           try {
@@ -141,6 +144,9 @@ router.post(
       );
 
       if (result.errors.length > 0) {
+        /**
+         * @type {Error & { info?: string }}
+         */
         const err = error.make(409, 'Failed to grant access to some users');
         err.info = '';
         const given_cp_and_cip = result.given_cp.filter(