+++ apps/prairielearn/src/tests/courseEditor.test.js
@@ -2,7 +2,7 @@
 const ERR = require('async-stacktrace');
 import { assert } from 'chai';
 import * as fs from 'fs-extra';
-const path = require('path');
+import * as path from 'path';
 import * as async from 'async';
 import * as cheerio from 'cheerio';
 import { exec } from 'child_process';