+++ tools/cjs-to-esm-candidates.mjs
@@ -7,11 +7,18 @@ import { parse } from '@typescript-eslint/parser';
 // These modules have their types declared as `export = ...`, so we can't use
 // `import` to load them until we're using native ESM.
 const CJS_ONLY_MODULES = new Set([
+  'archiver',
   'assert',
   'async-stacktrace',
   'axe-core',
+  'body-parser',
+  'byline',
   'cookie-parser',
+  'crypto-js/sha256',
   'csvtojson',
+  'detect-mocha',
+  'dockerode',
+  'events',
   'execa',
   'express',
   'express-async-handler',
@@ -19,13 +26,17 @@ const CJS_ONLY_MODULES = new Set([
   'fetch-cookie',
   'form-data',
   'json-stable-stringify',
+  'json-stringify-safe',
   'klaw',
   'lodash',
   'loopbench',
+  'oauth-signature',
   'node:assert',
   'passport',
   'postgres-interval',
+  'qrcode-svg',
   'request',
+  'search-string',
   'strip-ansi',
   'winston',
   'winston-transport',
@@ -52,6 +63,7 @@ function maybeLogLocation(filePath, node, modulePath) {
 }
 
 const importEqualsOnly = process.argv.includes('--import-equals-only');
+const filesWithImportsOnly = process.argv.includes('--files-with-imports-only');
 
 const files = await globby(['apps/*/src/**/*.{js,ts}']);
 
@@ -62,6 +74,15 @@ for (const file of files.sort()) {
     loc: true,
     tokens: false,
   });
+
+  const fileHasImports = ast.body.some(
+    (node) =>
+      node.type === 'ImportDeclaration' ||
+      (node.type === 'TSImportEqualsDeclaration' &&
+        node.moduleReference.type === 'TSExternalModuleReference'),
+  );
+
+  if (filesWithImportsOnly && !fileHasImports) continue;
 
   ast.body.forEach((node) => {
     // Handle `require()` calls.