+++ packages/migrations/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @prairielearn/migrations
 
+## 2.0.13
+
+### Patch Changes
+
+- abfd5cc: Upgrade all JavaScript dependencies
+- Updated dependencies [abfd5cc]
+- Updated dependencies [3249e13]
+  - @prairielearn/named-locks@1.5.10
+  - @prairielearn/postgres@1.9.0
+  - @prairielearn/logger@1.0.11
+  - @prairielearn/error@1.0.12
+
 ## 2.0.12
 
 ### Patch Changes